Config: replace `minimum_cushion` with `margin_usage`.
I found the concept of maintaining a minimum cushion to be too hard to
reason about, and it didn't properly reflect margin usage in a way that
made sense to me.

Thus, I've replaced it with `margin_usage`, which allows you to specify
how much margin to use as a factor of the net liquidation value. I find
this to be much better.