Improved handling of max new contracts.
I've replaced `target.maximum_new_contracts` with
`target.maximum_new_contracts_percent`.

Instead of limiting by a fixed number of contracts, the maximum number
of new contracts is calculated as a percentage of the buying power that
ThetaGang uses. This should solve the problem with adjusting for high
and low dollar value priced tickers, and allow for sensible scaling
based on account balances.

The default value is 5% of buying power, which should be a reasonable
default for most people.