Minor tweak: when rolling, allow same exp.
Currently we don't allow rolling up/down at the same expiration.
However, there isn't a good reason to disallow this so long as the
strikes are different.

This change allows rolling up/down at the same expiration provided the
strike price is different. Additionally, we avoid rolling to the same
price at a different expiration date.