Price update: adjust prices after submission
Sometimes the spreads are too wide on certain contracts, so we don't get
fills unless the price is adjusted after the order is submitted and the
spreads get less wide. The simplest fix is to just reset the price to
the midpoint again after submitting the order, although this 1) doesn't
always guarantee the order will fill and 2) may occasionally result in
less than ideal fills, but usually you still get an improvement over
the bid or ask price, provided it moves in the right direction.