Specify write thresholds as sigma
Rather than specifying the writing thresholds as a percentage of price,
we can use historical data to calculate standard deviation, and then
express the threshold in terms of sigma (i.e., relative to the standard
deviation).

For example, with `write_threshold_sigma = 1.0`, we'll write contracts
when the daily change is more (or less) than 1 times the daily standard
deviation for the last 30 days.

The amount of history used to compute the stddev can be altered using
the `constants.daily_stddev_window` setting. Refer to thetagang.toml for
more details.