Increase strike safety factor when scanning chains
When no strike limit was specified, we would previously scan above/below
the current underlying price, plus/minus 2%, when looking for contracts
to either roll to or write. The reason for this is that sometimes a 50
delta contract is ITM, so we might miss out on some juice.

This just increases that safety factor from 2% (relative to the
underlying price) to 5%. The only downside of this is that if we're
trying to write contracts really far OTM (such as 0.1 delta, for
example) we might need to increase the number of strikes to load from
the chain.