Optionally only write calls on excess positions
If you want to hold a stock/ETF, you may wish to avoid writing any calls
at all on it, except for the positions in excess of your desired
allocation. This may be preferable when you're bullish and want to avoid
missing out on upside.

For this case, you can just set either `write_when.calls.excess_only =
true` for the entire portfolio, or `symbols.<symbol>.calls.excess_only =
true` on a per-symbol basis.